All js files are now modules.  String size is now a separate arg. When a compiled program prints, or passes the address and size as separate i32 arguments now.
