+++ index.html
@@ -1,30 +1,33 @@
 <!DOCTYPE html>
 <html lang=en-US>
-  <head>
-    <title>TouchScript</title>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
-    <link rel="stylesheet" type="text/css" href="main.css">
-    <script type="module" src="main.mjs"></script>
-  </head>
-  
-  <body>
-    <div id="runtime">
-      <p id="console-output"></p>
-      <!-- <input id="console-input"> -->
-    </div>
+	<head>
+		<title>TouchScript</title>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
+		<link rel="stylesheet" type="text/css" href="main.css">
+		<script type="module" src="ui.mjs"></script>
+		<!-- <script type="module" src="main.mjs"></script> -->
+	</head>
+	
+	<body>
+		<div id="runtime">
+			<p id="console-output"></p>
+			<!-- <input id="console-input"> -->
+		</div>
 
-    
-    <div id="editor">
-      <div id="menu"><button class="indentation"></button><button class="insert-line menu-item"></button><button class="delete menu-item"></button></div>
-      
-      <div id="FAB-menu">
-        <button id="menu-button" class="FAB"></button>
-        <button id="view-code-button" class="FAB FAB-small"></button>
-        <button id="load-button" class="FAB FAB-small"></button>
-      </div>
-    </div>
+		
+		<div id="editor">
+			<div id="menu"><button class="indentation"></button><button class="insert-line menu-item"></button><button class="delete menu-item"></button></div>
+			
+			<div id="FAB-menu">
+				<button id="menu-button" class="FAB"></button>
+				<button id="view-code-button" class="FAB FAB-small"></button>
+				<button id="download-button" class="FAB FAB-small"></button>
+				<button id="load-button" class="FAB FAB-small"></button>
+				<button id="new-button" class="FAB FAB-small"></button>
+			</div>
+		</div>
 
-    <div id="program-list"><h3 id="no-projects-message">No projects found</h3></div>
-  </body>
+		<div id="program-list"><h3 id="no-projects-message">No projects found</h3></div>
+	</body>
 </html>
